+++ druid/src/theme.rs
@@ -60,6 +60,13 @@ pub const BORDERED_WIDGET_HEIGHT: Key<f64> =
 
 pub const TEXTBOX_BORDER_RADIUS: Key<f64> = Key::new("org.linebender.druid.theme.textbox_radius");
 
+/// The default horizontal spacing between widgets.
+pub const WIDGET_PADDING_HORIZONTAL: Key<f64> =
+    Key::new("org.linebender.druid.theme.widget-padding-h");
+/// The default vertical spacing between widgets.
+pub const WIDGET_PADDING_VERTICAL: Key<f64> =
+    Key::new("org.linebender.druid.theme.widget-padding-v");
+
 pub const SCROLLBAR_COLOR: Key<Color> = Key::new("org.linebender.druid.theme.scrollbar_color");
 pub const SCROLLBAR_BORDER_COLOR: Key<Color> =
     Key::new("org.linebender.druid.theme.scrollbar_border_color");
@@ -109,6 +116,8 @@ pub fn init() -> Env {
         .adding(SCROLLBAR_PAD, 2.)
         .adding(SCROLLBAR_RADIUS, 5.)
         .adding(SCROLLBAR_EDGE_WIDTH, 1.)
+        .adding(WIDGET_PADDING_VERTICAL, 10.0)
+        .adding(WIDGET_PADDING_HORIZONTAL, 8.0)
         .adding(
             UI_FONT,
             FontDescriptor::new(FontFamily::SYSTEM_UI).with_size(15.0),